requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-huggingface
langgraph
chromadb
pydantic
sentence-transformers
python-dotenv
tiktoken
tf-keras
langchain-google-genai
